pyproject.toml
[project]
name = "trueminds-deliverable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.129.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "resend>=2.21.0",
    "sqlmodel>=0.0.33",
    "uvicorn>=0.40.0",
]
